+++ tests/data/gradient.txt
@@ -1,26 +1,26 @@
--4.152456418715644304e-03 1.794258613907473432e-04 2.945898822104275894e-04
-7.583984572845370309e-03 -1.582973799291176320e-02 1.003090393109085099e-02
-7.108269288458313184e-03 -7.497428347602586030e-04 -1.832064845302960876e-02
-6.656100686175392411e-03 1.661920952938933382e-02 8.560363799203773993e-03
-4.150171297179205965e-03 -2.886184031214433787e-03 -1.928517966154341678e-03
--7.107469062184179620e-03 9.215403845574954289e-04 1.847625633155658420e-02
--7.447069365895675475e-03 1.627773778933860882e-02 -9.887612285597313644e-03
--4.695744153515590824e-03 6.617990860047218149e-03 1.333895988664401955e-03
-8.184225138430090141e-03 -1.908283627505280577e-02 1.124331267227107745e-02
-7.601018737479758869e-03 -1.453066307000039359e-03 -1.989289936360494337e-02
-3.476531273971517066e-03 -2.536134000308276626e-03 1.398261867402439407e-04
--7.146002269642348637e-03 1.188541316476943166e-03 1.890642484403404211e-02
--6.989945084689974664e-03 1.685085810027020514e-02 -1.030639604049365943e-02
--2.936584901913619206e-03 7.722262415632563659e-04 8.486175570587839445e-04
-6.217827017984678406e-03 -1.496893087649109333e-02 9.244588942414200261e-03
-6.794796096887267822e-03 -1.316171923393394394e-03 -1.882667934438586599e-02
-2.959560730930732575e-03 -5.854875217981003743e-03 5.540580172436193607e-04
--6.641073555246901350e-03 1.432979800375377977e-03 1.776945387166328771e-02
--7.504997483476216757e-03 1.794584732789372694e-02 -1.069944181114612647e-02
--1.283918493188984655e-03 3.654539406650059163e-03 2.298900813324511461e-03
-6.961339327204996152e-03 -1.606549022842296282e-02 9.831130518278552444e-03
-6.963669911264069867e-03 -1.290895391414803584e-03 -1.933202142367917986e-02
-2.399689404686267125e-03 2.122171025298325020e-03 -9.721940991136435173e-04
--7.394770889790347468e-03 7.634246340648687124e-04 1.905188256406953329e-02
--6.645974361674245046e-03 1.432153432873173315e-02 -9.191481784853738901e-03
--7.111902078638099312e-03 -1.763473892413269795e-02 -9.226691300157074949e-03
+-4.152457508121987217e-03 1.794440504521155333e-04 2.945870292148100589e-04
+7.583976070757447729e-03 -1.582974649555479729e-02 1.003091245260767650e-02
+7.108268696431885303e-03 -7.497371420916998375e-04 -1.832066611695548630e-02
+6.656096135093658717e-03 1.661921832679658095e-02 8.560369467724981973e-03
+4.150162737914797617e-03 -2.886190824669121469e-03 -1.928511665361121175e-03
+-7.107462233479910907e-03 9.215423882324991212e-04 1.847624721107443690e-02
+-7.447070473343142538e-03 1.627772521828330099e-02 -9.887607148872934459e-03
+-4.695738174131935949e-03 6.617986859358548912e-03 1.333897679256512703e-03
+8.184225717688953239e-03 -1.908282263679561552e-02 1.124332407370642883e-02
+7.601026709991298702e-03 -1.453064075729315618e-03 -1.989290990017655858e-02
+3.476546932557056380e-03 -2.536121200269469966e-03 1.398114168882358399e-04
+-7.146009121383745111e-03 1.188528518936138312e-03 1.890642031071587681e-02
+-6.989957887226783129e-03 1.685085072145042773e-02 -1.030639490556817250e-02
+-2.936578644141540906e-03 7.722271261334512360e-04 8.485816987979788451e-04
+6.217828443788597781e-03 -1.496892095276258772e-02 9.244591772095134274e-03
+6.794802923371090486e-03 -1.316175350374315656e-03 -1.882667991837116972e-02
+2.959557588722017130e-03 -5.854875232691458820e-03 5.540688338689925274e-04
+-6.641078369173936125e-03 1.432983787186259406e-03 1.776945330600465667e-02
+-7.505002311281039340e-03 1.794583881747913168e-02 -1.069944582182680293e-02
+-1.283902856530350078e-03 3.654545096820616124e-03 2.298913333587115915e-03
+6.961335379529476342e-03 -1.606550473487455832e-02 9.831114879677027574e-03
+6.963669658688331765e-03 -1.290899366290787498e-03 -1.933202228243668941e-02
+2.399691682586357899e-03 2.122184953878836211e-03 -9.722151481095231418e-04
+-7.394793093140616946e-03 7.634206650175556774e-04 1.905189963347098114e-02
+-6.645957585649231447e-03 1.432153774738598173e-02 -9.191489461768398428e-03
+-7.111899789080666778e-03 -1.763473637728107946e-02 -9.226679330565090709e-03
