+++ tests/data/gradient_hf.txt
@@ -1,26 +1,26 @@
-8.978915388979658019e-04 3.363908263054327108e-05 3.319854613437200896e-04
--7.038437388473539613e-05 1.794992954273766372e-04 -3.250090685469331220e-04
--5.866877472726628184e-04 -4.333401193434838206e-05 7.348526764872076456e-04
--4.777226081387198064e-04 -4.709471568276057951e-04 -3.675251036838744767e-04
-8.746846447138523217e-04 1.493301908794819610e-03 2.009144850489974488e-04
-8.166495035766274668e-05 -5.079289167042588815e-04 -9.595805749995633960e-04
--2.663063014551880769e-04 -4.247451101946353447e-04 1.273556904379802290e-04
--1.349621356894203927e-03 2.401833967269340064e-03 -9.733784622767416295e-04
-1.040448580624797614e-03 -2.431990253626281628e-03 1.221118566263612593e-03
-4.216678917146055028e-04 -1.154717210738454014e-04 -4.127608121962111909e-04
-4.834532774111721665e-07 2.072915208373160567e-03 2.739515139182913117e-03
--3.102513801422901452e-06 -2.599165327410446480e-04 -6.360716042763669975e-04
-2.567153956078982446e-04 6.698542165395338088e-05 -3.504496248751820531e-04
-6.112566097726812586e-04 -3.891267112976493081e-03 -1.683683592546003638e-03
--1.033286309848335804e-03 1.913489029448101064e-03 -7.456080197698611300e-04
--4.341822545939066913e-04 1.045006570166151505e-04 6.863993462502548937e-04
--6.083723746996838599e-04 -1.613954339063639054e-03 3.055949946428881958e-03
-5.743095243815332651e-04 8.435506515525048599e-05 -1.808193985652906122e-03
--2.966618239952367730e-04 1.240050835349393310e-03 -7.050030603750201408e-04
-2.452387661833199672e-03 -5.689673798769945279e-04 1.274290184483106714e-04
--3.066230647164047696e-04 6.246673933674173895e-04 -1.747565647747251205e-04
--3.402786319384176750e-04 7.855362582631641999e-05 3.900538558809785172e-05
--2.880164799989870517e-03 2.374506777869100915e-03 -9.598696237844706047e-04
-2.986698177664948162e-04 3.982592033935361542e-05 2.488178552084718831e-05
-1.078596767456474481e-03 -1.772169184732774738e-03 1.152215247657295549e-03
-6.461505108745768666e-05 -6.074890279705869034e-04 -3.396996817173203453e-04
+8.978904020295885857e-04 3.365727252457872964e-05 3.319826191727770492e-04
+-7.039290039756451733e-05 1.794907689145475160e-04 -3.250005420341040008e-04
+-5.866883157068514265e-04 -4.332832759246230125e-05 7.348350550273607951e-04
+-4.777271556122286711e-04 -4.709383460976823699e-04 -3.675194193419883959e-04
+8.746761182010232005e-04 1.493295087584556313e-03 2.009207378250721376e-04
+8.167177156792604364e-05 -5.079269271845987532e-04 -9.595896699465811253e-04
+-2.663074383235652931e-04 -4.247576157467847224e-04 1.273608063456777018e-04
+-1.349615388335223543e-03 2.401829988230019808e-03 -9.733767569741758052e-04
+1.040449149058986222e-03 -2.431976611205755034e-03 1.221129934947384754e-03
+4.216758497932460159e-04 -1.154694473370909691e-04 -4.127713282287004404e-04
+4.990852175978943706e-07 2.072927998142404249e-03 2.739500359894009307e-03
+-3.109335011686198413e-06 -2.599293225102883298e-04 -6.360761517498758622e-04
+2.567026058386545628e-04 6.697803200950147584e-05 -3.504484880068048369e-04
+6.112628625487559475e-04 -3.891266260325210169e-03 -1.683719403899885947e-03
+-1.033284888762864284e-03 1.913498977046401706e-03 -7.456051775989180896e-04
+-4.341754333836433943e-04 1.044972464114835020e-04 6.863987778160662856e-04
+-6.083755010877212044e-04 -1.613954339063639054e-03 3.055960746678465512e-03
+5.743046926909300964e-04 8.435904419457074255e-05 -1.808194554087094730e-03
+-2.966666556858399417e-04 1.240042308836564189e-03 -7.050070394143403973e-04
+2.452403293773386395e-03 -5.689616955351084471e-04 1.274415240004600491e-04
+-3.066270437557250261e-04 6.246528982956078835e-04 -1.747721967149118427e-04
+-3.402789161555119790e-04 7.854964678699616343e-05 3.900453293681493960e-05
+-2.880162526253116084e-03 2.374520704506721813e-03 -9.598906558494491037e-04
+2.986476488331391010e-04 3.982194130003335886e-05 2.489883854650543071e-05
+1.078613536265038420e-03 -1.772165774127643090e-03 1.152207573795749340e-03
+6.461732482421211898e-05 -6.074864700167381670e-04 -3.396877445993595757e-04
